pyproject.toml
[tool.poetry]
name = "user_management_svc"
version = "0.1.0-beta"
description = ""
authors = ["Mohammad Naim Faizy <mnaim.faizy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
fastapi = "^0.78.0"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.9.1"}
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
tenacity = "^8.0.1"
greenlet = "^1.1.2"
httpx = "^0.23.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cffi = "^1.15.0"
emails = "^0.6"
celery = "^5.2.7"
raven = "^6.10.0"
pyodbc = "^4.0.32"
fastapi-pagination = "^0.9.3"
orjson = "^3.7.6"
flower = "^1.1.0"
pika = "^1.3.0"
gunicorn = "^20.1.0"
python-crontab = "^2.6.0"
python-json-logger = "^2.0.4"
pandas = "^1.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^5.0.4"
black = {version = "^22.3.0", allow-prereleases = true}
autoflake = "^1.5.3"
isort = "^5.10.1"
sqlalchemy-stubs = "^0.4"
windows-curses = "^2.3.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
